+++ components/common/page-layout/Header.tsx
@@ -1,5 +1,4 @@
 import { useState } from 'react';
-import { CopyToClipboard } from 'react-copy-to-clipboard';
 import { useTheme } from '@emotion/react';
 import styled from '@emotion/styled';
 import Brightness4Icon from '@mui/icons-material/Brightness4';
@@ -9,20 +8,17 @@ import FavoritedIcon from '@mui/icons-material/Star';
 import NotFavoritedIcon from '@mui/icons-material/StarBorder';
 import { Box, CircularProgress } from '@mui/material';
 import IconButton from '@mui/material/IconButton';
-import FormControlLabel from '@mui/material/FormControlLabel';
 import Toolbar from '@mui/material/Toolbar';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
 import charmClient from 'charmClient';
-import Stack from '@mui/material/Stack';
-import Switch from '@mui/material/Switch';
-import LinkIcon from '@mui/icons-material/Link';
 import { useColorMode } from 'context/color-mode';
 import { usePages } from 'hooks/usePages';
 import { usePageTitle } from 'hooks/usePageTitle';
 import { useUser } from 'hooks/useUser';
 import { useCurrentSpace } from 'hooks/useCurrentSpace';
 import { useRouter } from 'next/router';
+import ShareButton from './ShareButton';
 
 export const headerHeight = 56;
 
@@ -57,11 +53,6 @@ export default function Header ({ open, openSidebar }: { open: boolean, openSide
       ? await charmClient.unfavoritePage(pageId)
       : await charmClient.favoritePage(pageId);
     setUser(newUser);
-  }
-
-  async function togglePublic (newPublicStatus: boolean) {
-    const updatedPage = await charmClient.togglePagePublicAccess(currentPage!.id, newPublicStatus);
-    setIsPublic(updatedPage.isPublic);
   }
 
   function generateShareLink () {
@@ -118,46 +109,10 @@ export default function Header ({ open, openSidebar }: { open: boolean, openSide
             </Box>
           )}
         </Box>
-        {/** favorite toggle */}
         <Box>
           {isPage && (
           <>
-            {
-            isAdmin === true && (
-              <>
-                <Tooltip title={isPublic ? 'Make private' : 'Make public'} arrow placement='bottom'>
-                  <FormControlLabel
-                    control={(
-                      <Switch
-                        defaultChecked={isPublic}
-                        onChange={ev => {
-                          togglePublic(ev.target.checked);
-                        }}
-                        inputProps={{ 'aria-label': 'toggle public access' }}
-                      />
-                )}
-                    label={isPublic === true ? 'Public' : 'Private'}
-                  />
-
-                </Tooltip>
-                {
-              isPublic === true && (
-                <Tooltip title='Copy sharing link' arrow placement='bottom'>
-
-                  <IconButton
-                    sx={{ ml: 1 }}
-                    color='inherit'
-                    onClick={generateShareLink}
-                  >
-                    {isPublic ? <LinkIcon color='secondary' /> : <LinkIcon color='secondary' />}
-                  </IconButton>
-
-                </Tooltip>
-              )
-            }
-              </>
-            )
-          }
+            {currentPage && <ShareButton page={currentPage} />}
 
             <Tooltip title={isFavorite ? 'Remove from Favorites' : 'Add to Favorites'} arrow placement='bottom'>
               <IconButton sx={{ ml: 1 }} onClick={toggleFavorite} color='inherit'>
